Add animation to button group
